Fixing checkstyle errors and moving event to aligning with other event location.
