+++ Essentials/src/main/java/com/earth2me/essentials/signs/SignSell.java
@@ -4,7 +4,7 @@ import com.earth2me.essentials.ChargeException;
 import com.earth2me.essentials.Trade;
 import com.earth2me.essentials.Trade.OverflowType;
 import com.earth2me.essentials.User;
-import com.earth2me.essentials.signs.event.SignTransactionEvent;
+import net.ess3.api.events.SignTransactionEvent;
 import net.ess3.api.IEssentials;
 import net.ess3.api.MaxMoneyException;
 import org.bukkit.inventory.ItemStack;
@@ -49,11 +49,14 @@ public class SignSell extends EssentialsSign {
 
         charge.isAffordableFor(player);
 
-        SignTransactionEvent signTransactionEvent = new SignTransactionEvent(
+        final SignTransactionEvent signTransactionEvent = new SignTransactionEvent(
+                sign,
+                this,
+                player,
                 charge.getItemStack(),
-                player.getBase(),
-                sign.getBlock().getLocation(),
-                SignTransactionEvent.TransactionType.SELL
+                SignTransactionEvent.TransactionType.SELL,
+                money.getMoney()
+
         );
         ess.getServer().getPluginManager().callEvent(signTransactionEvent);
         if (signTransactionEvent.isCancelled()) {
